+++ tpp/tensor_compare_lib.cpp
@@ -2,9 +2,9 @@
 #define _TENSOR_COMPARE_LIB_TPP_
 
 template <typename T>
-bool tensor_lib::compareArray(const T* A, const T* B, const unsigned Size)
+bool tensor_lib::compareArray(const T* A, const T* B, const size_t Size)
 {
-  for(unsigned i=0; i<Size; i++)
+  for(size_t i=0; i<Size; i++)
     if(A[i] != B[i])
       return false;
 
@@ -12,10 +12,10 @@ bool tensor_lib::compareArray(const T* A, const T* B, const unsigned Size)
 }
 
 template <typename T>
-bool tensor_lib::compare2Darray(const T** A, const T** B, const unsigned SizeY, const unsigned SizeX)
+bool tensor_lib::compare2Darray(const T** A, const T** B, const size_t SizeY, const size_t SizeX)
 {
-  for(unsigned i=0; i<SizeY; i++)
-    for(unsigned j=0; j<SizeX; j++)
+  for(size_t i=0; i<SizeY; i++)
+    for(size_t j=0; j<SizeX; j++)
       if(A[i][j] != B[i][j])
         return false;
 
@@ -25,11 +25,11 @@ bool tensor_lib::compare2Darray(const T** A, const T** B, const unsigned SizeY, 
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<T> & A, const std::vector<T> & B)
 {
-  unsigned C = A.size();
+  size_t C = A.size();
 
   if(C != B.size()) return false;
 
-  for(unsigned c=0; c<C; c++)
+  for(size_t c=0; c<C; c++)
     if(A[c] != B[c])
       return false;
 
@@ -39,14 +39,14 @@ bool tensor_lib::compareTensors(const std::vector<T> & A, const std::vector<T> &
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<T> >& A,const std::vector<std::vector<T> >& B)
 {
-  unsigned H = A.size();
-  unsigned W = A[0].size();
+  size_t H = A.size();
+  size_t W = A[0].size();
 
   if(H != B.size()) return false;
   if(W != B[0].size()) return false;
 
-  for(unsigned h=0; h<H; h++)
-    for(unsigned w=0; w<W; w++)
+  for(size_t h=0; h<H; h++)
+    for(size_t w=0; w<W; w++)
       if(A[h][w] != B[h][w])
         return false;
 
@@ -56,17 +56,17 @@ bool tensor_lib::compareTensors(const std::vector<std::vector<T> >& A,const std:
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<T> > >& A,const std::vector<std::vector<std::vector<T> > >& B)
 {
-  unsigned C = A.size();
-  unsigned H = A[0].size();
-  unsigned W = A[0][0].size();
+  size_t C = A.size();
+  size_t H = A[0].size();
+  size_t W = A[0][0].size();
 
   if(C != B.size()) return false;
   if(H != B[0].size()) return false;
   if(W != B[0][0].size()) return false;
 
-  for(unsigned c=0; c<C; c++)
-    for(unsigned h=0; h<H; h++)
-      for(unsigned w=0; w<W; w++)
+  for(size_t c=0; c<C; c++)
+    for(size_t h=0; h<H; h++)
+      for(size_t w=0; w<W; w++)
         if(A[c][h][w] != B[c][h][w])
           return false;
 
@@ -76,20 +76,20 @@ bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<T> > >
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<std::vector<T> > > >& A,const std::vector<std::vector<std::vector<std::vector<T> > > >& B)
 {
-  unsigned K = A.size();
-  unsigned C = A[0].size();
-  unsigned H = A[0][0].size();
-  unsigned W = A[0][0][0].size();
+  size_t K = A.size();
+  size_t C = A[0].size();
+  size_t H = A[0][0].size();
+  size_t W = A[0][0][0].size();
 
   if(K != B.size()) return false;
   if(C != B[0].size()) return false;
   if(H != B[0][0].size()) return false;
   if(W != B[0][0][0].size()) return false;
 
-  for(unsigned k=0; k<K; k++)
-    for(unsigned c=0; c<C; c++)
-      for(unsigned h=0; h<H; h++)
-        for(unsigned w=0; w<W; w++)
+  for(size_t k=0; k<K; k++)
+    for(size_t c=0; c<C; c++)
+      for(size_t h=0; h<H; h++)
+        for(size_t w=0; w<W; w++)
           if(A[k][c][h][w] != B[k][c][h][w])
             return false;
 
@@ -98,9 +98,9 @@ bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<std::v
 
 /////////////////////////////////////
 template <typename T>
-bool tensor_lib::compareArray(const T* A, const T* B, const unsigned Size, const T t)
+bool tensor_lib::compareArray(const T* A, const T* B, const size_t Size, const T t)
 {
-  for(unsigned i=0; i<Size; i++)
+  for(size_t i=0; i<Size; i++)
     if(std::abs((A[i] - B[i]) / B[i]) > t)
       return false;
 
@@ -108,10 +108,10 @@ bool tensor_lib::compareArray(const T* A, const T* B, const unsigned Size, const
 }
 
 template <typename T>
-bool tensor_lib::compare2Darray(const T** A, const T** B, const unsigned SizeY, const unsigned SizeX, const T t)
+bool tensor_lib::compare2Darray(const T** A, const T** B, const size_t SizeY, const size_t SizeX, const T t)
 {
-  for(unsigned i=0; i<SizeY; i++)
-    for(unsigned j=0; j<SizeX; j++)
+  for(size_t i=0; i<SizeY; i++)
+    for(size_t j=0; j<SizeX; j++)
       if(std::abs((A[i][j] - B[i][j]) / B[i][j]) > t)
         return false;
 
@@ -121,11 +121,11 @@ bool tensor_lib::compare2Darray(const T** A, const T** B, const unsigned SizeY, 
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<T> & A,const std::vector<T> & B,const T t)
 {
-  unsigned C = A.size();
+  size_t C = A.size();
 
   if(C != B.size()) return false;
 
-  for(unsigned c=0; c<C; c++)
+  for(size_t c=0; c<C; c++)
     if(std::abs((A[c] - B[c]) / B[c]) > t)
       return false;
 
@@ -135,14 +135,14 @@ bool tensor_lib::compareTensors(const std::vector<T> & A,const std::vector<T> & 
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<T> >& A,const std::vector<std::vector<T> >& B,const T t)
 {
-  unsigned H = A.size();
-  unsigned W = A[0].size();
+  size_t H = A.size();
+  size_t W = A[0].size();
 
   if(H != B.size()) return false;
   if(W != B[0].size()) return false;
 
-  for(unsigned h=0; h<H; h++)
-    for(unsigned w=0; w<W; w++)
+  for(size_t h=0; h<H; h++)
+    for(size_t w=0; w<W; w++)
       if(std::abs((A[h][w] - B[h][w]) / B[h][w]) > t)
         return false;
 
@@ -152,17 +152,17 @@ bool tensor_lib::compareTensors(const std::vector<std::vector<T> >& A,const std:
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<T> > >& A,const std::vector<std::vector<std::vector<T> > >& B,const T t)
 {
-  unsigned C = A.size();
-  unsigned H = A[0].size();
-  unsigned W = A[0][0].size();
+  size_t C = A.size();
+  size_t H = A[0].size();
+  size_t W = A[0][0].size();
 
   if(C != B.size()) return false;
   if(H != B[0].size()) return false;
   if(W != B[0][0].size()) return false;
 
-  for(unsigned c=0; c<C; c++)
-    for(unsigned h=0; h<H; h++)
-      for(unsigned w=0; w<W; w++)
+  for(size_t c=0; c<C; c++)
+    for(size_t h=0; h<H; h++)
+      for(size_t w=0; w<W; w++)
         if(std::abs((A[c][h][w] - B[c][h][w]) / B[c][h][w]) > t)
           return false;
 
@@ -172,20 +172,20 @@ bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<T> > >
 template <typename T>
 bool tensor_lib::compareTensors(const std::vector<std::vector<std::vector<std::vector<T> > > >& A,const std::vector<std::vector<std::vector<std::vector<T> > > >& B,const T t)
 {
-  unsigned K = A.size();
-  unsigned C = A[0].size();
-  unsigned H = A[0][0].size();
-  unsigned W = A[0][0][0].size();
+  size_t K = A.size();
+  size_t C = A[0].size();
+  size_t H = A[0][0].size();
+  size_t W = A[0][0][0].size();
 
   if(K != B.size()) return false;
   if(C != B[0].size()) return false;
   if(H != B[0][0].size()) return false;
   if(W != B[0][0][0].size()) return false;
 
-  for(unsigned k=0; k<K; k++)
-    for(unsigned c=0; c<C; c++)
-      for(unsigned h=0; h<H; h++)
-        for(unsigned w=0; w<W; w++)
+  for(size_t k=0; k<K; k++)
+    for(size_t c=0; c<C; c++)
+      for(size_t h=0; h<H; h++)
+        for(size_t w=0; w<W; w++)
           if(std::abs((A[k][c][h][w] - B[k][c][h][w]) / B[k][c][h][w]) > t)
             return false;
